new file structure, changes in tests
